Fixed merge conflict and minor improvement
